% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_hex_grids.R
\docType{data}
\name{ca_hex_grid_res_4}
\alias{ca_hex_grid_res_4}
\alias{cr_hex_grid_res_5}
\alias{cr_hex_grid_res_6}
\alias{cr_hex_grid_res_7}
\alias{cr_hex_grid_res_8}
\title{Hierarchical hexagonal geospatial grids for Central America of resolution 4, 5, 6, 7 and 8 (H3, Hexagonal hierarchical geospatial indexing system)}
\format{
An sf object (EPSG:4326)
}
\usage{
ca_hex_grid_res_4
ca_hex_grid_res_5
ca_hex_grid_res_6
ca_hex_grid_res_7
ca_hex_grid_res_8
}
\description{
Hierarchical hexagonal geospatial grids for Central America of resolution 4, 5, 6, 7 and 8 (H3, Hexagonal hierarchical geospatial indexing system).
}
\keyword{datasets}
